Commit before trying new method
This is currently set up to dynamically add new contacts to an html table with buttons to edit or delete items on the row.
The difficulty is how to access this row data and process the edit/delete functionality without alot of javascript/jquery.  I am checking this in before I try another path of using CKEditor hbox and dynamically adding records to when the add button is clicked.